+++ onprem/core.py
@@ -7,12 +7,13 @@ __all__ = ['DEFAULT_MODEL_URL', 'DEFAULT_LARGER_URL', 'DEFAULT_EMBEDDING_MODEL',
 # %% ../nbs/00_core.ipynb 3
 from langchain.chains import RetrievalQA, ConversationalRetrievalChain
 from langchain.memory import ConversationBufferMemory
-from langchain.embeddings import HuggingFaceEmbeddings
 from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
 from langchain.callbacks.manager import CallbackManager
-from langchain.vectorstores import Chroma
-from langchain.llms import LlamaCpp
 from langchain.prompts import PromptTemplate
+from langchain_community.embeddings import HuggingFaceEmbeddings
+from langchain_community.vectorstores import Chroma
+from langchain_community.llms import LlamaCpp
+from langchain_openai import ChatOpenAI
 import chromadb
 import os
 import warnings
@@ -46,11 +47,13 @@ class LLM:
         model_url=DEFAULT_MODEL_URL,
         use_larger: bool = False,
         n_gpu_layers: Optional[int] = None,
+        prompt_template: Optional[str] = None,
         model_download_path: Optional[str] = None,
         vectordb_path: Optional[str] = None,
         max_tokens: int = 512,
-        n_ctx: int = 2048,
+        n_ctx: int = 3900,
         n_batch: int = 1024,
+        stop:list=[],
         mute_stream: bool = False,
         callbacks=[],
         embedding_model_name: str = "sentence-transformers/all-MiniLM-L6-v2",
@@ -67,16 +70,19 @@ class LLM:
 
         **Args:**
 
-        - *model_url*: URL to `.bin` model (currently must be GGML model).
-                       Can also be the filename of model if it has already been downloaded to `model_download_path`.
+        - *model_url*: URL to `.GGUF` model (or the filename if already been downloaded to `model_download_path`).
+                       To use a non-local OpenAI model instead, replace URL with: `openai://<name_of_model>` 
+                       (e.g., `openai://gpt-3.5-turbo`).
         - *use_larger*: If True, a larger model than the default `model_url` will be used.
         - *n_gpu_layers*: Number of layers to be loaded into gpu memory. Default is `None`.
+        - *prompt_template*: Optional prompt template (must have a variable named "prompt").
         - *model_download_path*: Path to download model. Default is `onprem_data` in user's home directory.
         - *vectordb_path*: Path to vector database (created if it doesn't exist).
                            Default is `onprem_data/vectordb` in user's home directory.
         - *max_tokens*: The maximum number of tokens to generate.
-        - *n_ctx*: Token context window.
+        - *n_ctx*: Token context window. (Llama2 models have max of 4096.)
         - *n_batch*: Number of tokens to process in parallel.
+        - *stop*: a list of strings to stop generation when encountered (applied to all calls to `LLM.prompt`)
         - *mute_stream*: Mute ChatGPT-like token stream output during generation
         - *callbacks*: Callbacks to supply model
         - *embedding_model_name*: name of sentence-transformers model. Used for `LLM.ingest` and `LLM.ask`.
@@ -91,12 +97,13 @@ class LLM:
         self.model_url = DEFAULT_LARGER_URL if use_larger else model_url
         self.model_name = os.path.basename(self.model_url)
         self.model_download_path = model_download_path or U.get_datadir()
-        if not os.path.isfile(os.path.join(self.model_download_path, self.model_name)):
+        if not os.path.isfile(os.path.join(self.model_download_path, self.model_name)) and not self.is_openai_model():
             self.download_model(
                 self.model_url,
                 model_download_path=self.model_download_path,
                 confirm=confirm,
             )
+        self.prompt_template = prompt_template
         self.vectordb_path = vectordb_path
         self.llm = None
         self.ingester = None
@@ -106,6 +113,8 @@ class LLM:
         self.max_tokens = max_tokens
         self.n_ctx = n_ctx
         self.n_batch = n_batch
+        self.stop = stop
+        self.mute_stream = mute_stream
         self.callbacks = [] if mute_stream else [StreamingStdOutCallbackHandler()]
         if callbacks:
             self.callbacks.extend(callbacks)
@@ -116,7 +125,38 @@ class LLM:
         self.rag_score_threshold = rag_score_threshold
         self.verbose = verbose
         self.extra_kwargs = kwargs
+
+
+        # explicitly set offload_kqv
+        # reference: https://github.com/abetlen/llama-cpp-python/issues/999#issuecomment-1858041458
+        self.offload_kqv = True if n_gpu_layers is not None and n_gpu_layers > 0 else False
+
+        # load LLM
         self.load_llm()
+
+        # issue warning
+        if self.is_openai_model():
+            warnings.warn(f'The model you supplied is {self.model_name}, an external service (i.e., not on-premises). '+\
+                          f'Use with caution, as your data and prompts will be sent externally.')
+
+    def is_openai_model(self):
+        return self.model_url.lower().startswith('openai')
+
+    def update_max_tokens(self, value:int=512):
+        """
+        Update `max_tokens` (maximum length of generation).
+        """
+        llm = self.load_llm()
+        llm.max_tokens = value
+
+
+    def update_stop(self, value:list=[]):
+        """
+        Update `max_tokens` (maximum length of generation).
+        """
+        llm = self.load_llm()
+        llm.stop = value
+
 
     @classmethod
     def download_model(
@@ -209,6 +249,7 @@ class LLM:
         """
         Returns the path to the model
         """
+        if self.is_openai_model(): return None
         datadir = self.model_download_path
         model_path = os.path.join(datadir, self.model_name)
         if not os.path.isfile(model_path):
@@ -224,8 +265,14 @@ class LLM:
         """
         model_path = self.check_model()
 
-        if not self.llm:
-            self.llm = llm = LlamaCpp(
+        if not self.llm and self.is_openai_model():
+            self.llm = ChatOpenAI(model_name=self.model_name, 
+                                  callback_manager=CallbackManager(self.callbacks), 
+                                  streaming=not self.mute_stream,
+                                  max_tokens=self.max_tokens,
+                                  **self.extra_kwargs)
+        elif not self.llm:
+            self.llm = LlamaCpp(
                 model_path=model_path,
                 max_tokens=self.max_tokens,
                 n_batch=self.n_batch,
@@ -233,24 +280,37 @@ class LLM:
                 verbose=self.verbose,
                 n_gpu_layers=self.n_gpu_layers,
                 n_ctx=self.n_ctx,
+                offload_kqv = self.offload_kqv,
                 **self.extra_kwargs,
             )
 
         return self.llm
 
-    def prompt(self, prompt, prompt_template: Optional[str] = None):
+
+    def prompt(self, prompt, prompt_template: Optional[str] = None, stop:list=[], **kwargs):
         """
-        Send prompt to LLM to generate a response
+        Send prompt to LLM to generate a response.
+        Extra keyword arguments are sent directly to the model invocation.
 
         **Args:**
 
         - *prompt*: The prompt to supply to the model
-        - *prompt_template*: Optional prompt template (must have a variable named "prompt")
+        - *prompt_template*: Optional prompt template (must have a variable named "prompt").
+                             This value will override any `prompt_template` value supplied 
+                             to `LLM` constructor.
+        - *stop*: a list of strings to stop generation when encountered. 
+                  This value will override the `stop` parameter supplied to `LLM` constructor.
+
         """
         llm = self.load_llm()
+        prompt_template = self.prompt_template if prompt_template is None else prompt_template
         if prompt_template:
             prompt = prompt_template.format(**{"prompt": prompt})
-        return llm(prompt)
+        stop = stop if stop else self.stop
+        res = llm.invoke(prompt, stop=stop, **kwargs)
+        return res.content if self.is_openai_model() else res
+
+
 
     def load_qa(self, prompt_template: str = DEFAULT_QA_PROMPT):
         """
@@ -304,31 +364,38 @@ class LLM:
             )
         return self.chatqa
 
-    def ask(self, question: str, prompt_template=DEFAULT_QA_PROMPT):
+    def ask(self, question: str, qa_template=DEFAULT_QA_PROMPT, prompt_template=None, **kwargs):
         """
         Answer a question based on source documents fed to the `ingest` method.
+        Extra keyword arguments are sent directly to the model invocation.
 
         **Args:**
 
         - *question*: a question you want to ask
-        - *prompt_template*: A string representing the prompt with variables "context" and "question"
+        - *qa_template*: A string representing the prompt with variables "context" and "question"
+        - *prompt_template*: the model-specific template in which everything (including QA template) should be wrapped.
+                            Should have a single variable "{prompt}". Overrides the `prompt_template` parameter supplied to 
+                            `LLM` constructor.
 
         **Returns:**
 
         - A dictionary with keys: `answer`, `source_documents`, `question`
         """
+        prompt_template = self.prompt_template if prompt_template is None else prompt_template
+        prompt_template = qa_template if prompt_template is None else prompt_template.format(**{'prompt': qa_template})
         qa = self.load_qa(prompt_template=prompt_template)
-        res = qa(question)
+        res = qa.invoke(question, **kwargs)
         res["question"] = res["query"]
         del res["query"]
         res["answer"] = res["result"]
         del res["result"]
         return res
 
-    def chat(self, question: str):
+    def chat(self, question: str, **kwargs):
         """
         Chat with documents fed to the `ingest` method.
         Unlike `LLM.ask`, `LLM.chat` includes conversational memory.
+        Extra keyword arguments are sent directly to the model invocation.
 
         **Args:**
 
@@ -339,5 +406,5 @@ class LLM:
         - A dictionary with keys: `answer`, `source_documents`, `question`, `chat_history`
         """
         chatqa = self.load_chatqa()
-        res = chatqa(question)
+        res = chatqa.invoke(question, **kwargs)
         return res